Updates the shuttle engine sprites. Renames HidrogenTubes to ShuttleHeater. Adds PipeTrigger to separate the Interact() from Pipe.cs. Adds a simple piping system for the ShuttleHeater.
