+++ UnityProject/Assets/Resources/Prefabs/Objects/Atmos/AirVent.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 61000013103678858}
   - component: {fileID: 114671962178843966}
   - component: {fileID: 4396274152352043513}
+  - component: {fileID: 6203799934882642808}
   - component: {fileID: 7499999529140733009}
   - component: {fileID: 8354343867630940287}
   - component: {fileID: 966013011358701611}
@@ -109,6 +110,18 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 579884b05ae74e8499329b64997147d3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &6203799934882642808
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1000010733523080}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 027d1f2fae780624ebb2115cfc6e78af, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &7499999529140733009
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -124,14 +137,12 @@ MonoBehaviour:
   nodes: []
   anchored: 0
   direction: 0
+  registerTile: {fileID: 0}
   pipeSprites:
   - {fileID: 21300224, guid: b17c2d8fde0016741b3ddae80bb61948, type: 3}
   - {fileID: 21300162, guid: 15ef494d1ff854fa69eab656aa860854, type: 3}
   spriteRenderer: {fileID: 8418407404781841632}
   volume: 70
-  ARANpipenetmembers: 0
-  ARANpipenetName: 
-  ARANpipenetVolume: 0
   MinimumPressure: 101.325
 --- !u!114 &8354343867630940287
 MonoBehaviour:
