+++ UnityProject/Assets/Resources/Prefabs/Objects/Atmos/Pipe.prefab
@@ -91,6 +91,7 @@ GameObject:
   - component: {fileID: 61321219257688126}
   - component: {fileID: 114348495363464762}
   - component: {fileID: 114675370128025870}
+  - component: {fileID: 5640039996929275955}
   - component: {fileID: 3760268221926563392}
   - component: {fileID: 114836435796936700}
   - component: {fileID: 114007797438959630}
@@ -187,6 +188,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 579884b05ae74e8499329b64997147d3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &5640039996929275955
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1645856876941636}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 027d1f2fae780624ebb2115cfc6e78af, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &3760268221926563392
