+++ UnityProject/Assets/Scripts/Shuttles/ShuttleHeater.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Atmospherics;
+
+public class ShuttleHeater : AdvancedPipe
+{
+	public override void SpriteChange()
+	{
+		return;
+	}
+}
