+++ UnityProject/Assets/Scripts/Shuttles/ShuttleHeater.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 352c75cf35ef70b4ba50e454d14aa536
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
